+++ CHANGES.md
@@ -22,12 +22,15 @@ Released: --
   - Add `Form` field and `form_and_files` location for better form upload support.
   - Rewrite the `files` to act like `form_and_files`, so that failed parsed file will
     be included in the returned data.
+- Allow to use `json_or_form` location and fields (`DelimitedList` and `DelimitedTuple`)
+  from webargs ([issue #254][issue_254]).
 
 [issue_211]: https://github.com/greyli/apiflask/issues/211
 [issue_231]: https://github.com/greyli/apiflask/issues/231
 [pull_232]: https://github.com/greyli/apiflask/pull/232
 [issue_237]: https://github.com/greyli/apiflask/issues/237
 [issue_123]: https://github.com/greyli/apiflask/issues/123
+[issue_254]: https://github.com/greyli/apiflask/issues/254
 
 
 ## Version 0.12.0